+++ lang/en/controllers/table.php
@@ -4,6 +4,7 @@ return [
     'columns' => [
         'id' => 'ID',
         'callsign' => 'Callsign',
+        'description' => 'Description',
         'frequency' => 'Frequency',
         'top_down' => 'Top-down for',
     ],