+++ lang/en/stands/table.php
@@ -2,10 +2,11 @@
 
 return [
     'columns' => [
-        'id' => 'Id',
         'airfield' => 'Airfield',
         'terminal' => 'Terminal',
         'identifier' => 'Identifier',
+        'max_wtc' => 'Max UK WTC',
+        'max_size' => 'Max Aircraft Size',
         'airlines' => 'Airlines',
         'used' => 'Used for Allocation',
         'priority' => 'Allocation Priority',